+++ projects/igniteui-angular/src/lib/directives/forms-control/forms-control.directive.ts
@@ -4,16 +4,12 @@ import {
     ElementRef,
     HostListener,
     NgModule,
-    Renderer2,
-    AfterViewInit,
-    OnDestroy
+    Renderer2
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { fromEvent, Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
 
 @Directive({
-    selector: 'igc-rating', /* eslint-disable-line @angular-eslint/directive-selector */
+    selector: 'igc-rating[ngModel],igc-rating[formControlName]', /* eslint-disable-line @angular-eslint/directive-selector */
     providers: [
         {
             provide: NG_VALUE_ACCESSOR,
@@ -22,47 +18,23 @@ import { takeUntil } from 'rxjs/operators';
         }
     ]
 })
-export class IgxFormsControlDirective implements ControlValueAccessor, AfterViewInit, OnDestroy {
+export class IgxFormsControlDirective implements ControlValueAccessor {
     /** @hidden @internal */
-    public onChange: any = () => { };
+    private onChange: any = () => { };
     /** @hidden @internal */
-    public onTouched: any = () => { };
-
-    protected _destroy$ = new Subject();
-
-    private _value: any;
-
-    /** @hidden @internal */
-    public get value(): any {
-        return this.elementRef.nativeElement.value;
-    }
-    public set value(val) {
-        if (val !== this._value) {
-            this._value = val;
-            this.onChange(this._value);
-            this.elementRef.nativeElement.value = val;
-        }
-    }
+    private onTouched: any = () => { };
 
     constructor(
         private elementRef: ElementRef,
         private renderer: Renderer2) { }
 
     /** @hidden @internal */
-    public ngAfterViewInit(): void {
-        fromEvent(this.elementRef.nativeElement, 'blur')
-            .pipe(takeUntil(this._destroy$))
-            .subscribe(() => {
-                this.onTouched();
-            });
+    @HostListener('blur')
+    public onBlur() {
+        this.onTouched();
     }
 
     /** @hidden @internal */
-    public ngOnDestroy(): void {
-        this._destroy$.next();
-        this._destroy$.complete();
-    }
-
     @HostListener('igcChange', ['$event.detail'])
     public listenForValueChange(value) {
         this.onChange(value);
@@ -71,7 +43,6 @@ export class IgxFormsControlDirective implements ControlValueAccessor, AfterView
     /** @hidden @internal */
     public writeValue(value): void {
         if (value) {
-            this._value = value;
             this.elementRef.nativeElement.value = value;
         }
     }